lv2 commit

+++ routes/comments.js
@@ -1,11 +1,12 @@
 //express 기능 사용
 const express = require('express');
 const router = express.Router();
+const authMiddleware = require("../middlewares/auth-middleware")
 
 //schemes 불러오기
 const Comments = require('../schemas/comment.js');
 
-router.route('/:_postId')
+router.route('/comments/:_postId')
     //댓글 목록 조회
     .get(async (req, res) => {
         try {
@@ -25,14 +26,15 @@ router.route('/:_postId')
         }
     });
 
-router.route('/:_postId')
+router.route('/comments/:_postId')
     //댓글 생성
-    .post(async (req, res) => {
+    .post(authMiddleware, async (req, res) => {
         try {
             const postId = req.params._postId
             try {
+                const {userId} = res.locals.user
                 const { user, password, content } = req.body
-                await Comments.create({ postId: postId, user, password, content })
+                await Comments.create({ postId: postId, user, password, content, userId })
                 return res.status(200).json({ message: '댓글을 생성하였습니다.' })
             } catch {
                 return res.status(400).json({ message: '댓글 내용을 입력해주세요.' });
@@ -42,19 +44,20 @@ router.route('/:_postId')
         }
     })
 
-router.route('/:_commentId')
+router.route('/comments/:_commentId')
 //댓글 수정
-    .put(async (req, res) => {
+    .put(authMiddleware, async (req, res) => {
         try {
+            const {userId} = res.locals.user
             const commentId = req.params._commentId
-            const comment = await Comments.findById(commentId)
+            const comment = await Comments.findById(userId, commentId)
             const {password, content} = req.body
 
             if(!comment) {
                 return res.status(404).json({message: '댓글 조회에 실패하였습니다.'})
             }
             if (password === comment.password) {
-                await Comments.updateOne({_id: commentId}, {$set: {content:content} })
+                await Comments.updateOne({_id: commentId}, {$set: {content:content, userId:userId}})
                 return res.status(200).json({message:'댓글을 수정하였습니다.'})
             } else {
                 return res.status(404).json
@@ -65,16 +68,17 @@ router.route('/:_commentId')
         }
     })
 //댓글 삭제
-    .delete(async (req,res) => {
+    .delete(authMiddleware, async (req,res) => {
         try {
+            const {userId} = res.locals.user
             const commentId = req.params._commentId
-            const comment = await Comments.findById(commentId)
+            const comment = await Comments.findById(userId, commentId)
             const password = req.body.password
             if (!comment) {
                 return res.status(404).json({message: '댓글 조회에 실패하였습니다.'})
             }
             if (password === comment.password) {
-                await Comments.deleteOne({_id: commentId})
+                await Comments.deleteOne({userId, _id: commentId})
                 return res.status(200).json({message: '댓글을 삭제하였습니다.'})
             } else {
                 return res.status(404).json
